Backend/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
openai==1.3.0
python-dotenv==1.0.0
googlemaps==4.10.0
python-dateutil==2.8.2
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0 